backend/requirements.txt
fastapi
sqlalchemy
bcrypt
pydantic
imaplib2
email-validator
PyPDF2
python-docx
pandas
pytesseract
python-dotenv
neo4j
uvicorn
psycopg2-binary
lightrag-hku
groq
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.0.2
tenacity==8.2.2
agno-agent==0.1.0
pydantic==2.5.0